Removing ES5 version
