+++ .gitignore
@@ -22,4 +22,3 @@ dev
 pub
 deploy
 lib/
-dist/
