+++ package.json
@@ -1,12 +1,13 @@
 {
   "name": "wtc-utility-helpers",
-  "version": "1.3.3",
+  "version": "2.0.0",
   "description": "Javascript helpers, perfect for non-jquery projects.",
   "main": "dist/wtc-utility-helpers.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "babel src/es6/ --presets babel-preset-es2015 --out-dir dist",
-    "prepublish": "npm run build"
+    "build": "babel src/ --presets @babel/preset-env --out-dir dist",
+    "build-ES5": "npx webpack --config webpack.config.js",
+    "prepublish": "npm run build && npm run build-ES5"
   },
   "repository": {
     "type": "git",
@@ -24,7 +25,16 @@
   },
   "homepage": "https://github.com/wethegit/wtc-utility-helpers",
   "devDependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "^7.2.3",
+    "babel": "^6.23.0",
     "babel-cli": "^6.18.0",
-    "babel-preset-es2015": "^6.18.0"
+    "babel-loader": "^8.0.5",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.18.0",
+    "webpack": "^4.28.1",
+    "webpack-cli": "^3.2.1"
   }
 }
